Add 2020/otterness/try.sh
The .gitignore was updated but it cannot be sorted because it excludes
'*.mid' but then negates the .mid files in the repo, cvikl.mid and
entertainer.mid. These have to come after the exclusion of '*.mid'. This
is important as one might use the entry to create a .mid file that is
not done in the try script. The make clobber rule does not remove
'*.mid' files as doing so would remove the ones we want to keep.