Add tmp/update-manifest.sh
This script helps speed up updating the manifest files. It prompts many
times to make sure everything is okay after the user (hopefully)
carefully looks at output. If any of the scripts return non-zero it
exits anyway.

The git commit commands CAN be automated but one MUST specify both the
'-a' option AND say 'y' (or 'Y') to the prompt. Otherwise it tells the
user what commands to run and then prompts them to press a key after
they did that. Sometimes it prompts more than once just to be sure.

It prompts if you want to use the quick-readme2index.sh script as well
but if anything but 'y' or 'Y' is specified it will use the slower
script.

The default value for -v for other tools is 3. This is because it shows
progress and it's easier to not have to remember to pass -D 3 (or
whatever). One may pass -D 0 to set back to 0 should they wish to.

So although it does help automate the process of updating the manifest
it is very interactive and it must remain that way.

I have tested the paths of the script: using the quick script,
automatically running git commands and also not running them
automatically (in which case one MUST run it in another shell). The
limitation is that the script does not let one choose which steps to do.
This is because to update the manifest properly one must follow all the
steps and when there is a problem one must start over from the
beginning.