Typo fixes and comments updated in top Makefile
To do with the build of html files in particular. The comments for each
now have the tool being used in the form of ${FOO} tool (bin/foo).

The entry_index rule has been aliased to readme2index as that is what
the rule uses (well bin/readme2index.sh).

The quick_entry_index rule has been aliased to quick_readme2index as
that is what the rule uses (well bin/quick-readme2index.sh).