Typo/format and other fixes to top README.md
'Other fixes' also include an expansion on some thoughts. For instance
it was suggested that to understand an entry one might change variable
and label names but some entries will actually break if you do this.
Another example is that some beautifiers will actually prevent code from
compiling. Another is that some lints crash (it was already stated
'abort' but that does not necessarily mean crash; splint for instance
will stop parsing quite easily if it doesn't like what it sees which
happens quite often). There probably were other things like this that
were updated/added/changed.

The format and typo fixes are less significant but I believe and hope it
should look better now (in particular the format fixes and adjustments).