Update 2012/zeitak test script + how to view code
This involves a couple java files being added to the repo both with a
joke included (to make it more fun).

The instructions of how to view the code are to make the layout correct
(at least for those who use vim (vi too? Been too long now)).

This commit will require a manifest update which will happen no later
than this weekend but it might happen sooner.