Update todo file
The task of weasel.spec was removed as it's complete.

Rather than have to open a link to find the links to check (many if not
all are probably done already but they have to be confirmed yet) at
issue 3 I have just put the direct links in the file.

Also removed the embed video task. It might be useful but it's probably
not worth delaying things more.