Add FAQ entry about opening pull requests
This used commit bbca4cb97f54feb378c53f2b82250bad7d5edd9b as an example
(an example commit and pull request was suggested to me by Landon).

Note that I did NOT update the html file as I am not clear how to do
this (for the moment).