Minor fixes and updates to FAQ
'FAQ 6.10: How does someone make a change to a file and submit that
change as a GitHub pull request?' has been updated to note that the
authors of winning entries tend to have more deference but that the
judges still have the final say in the matter.

There were some typos (including links) and formatting fixes as well.