remove XXX from text in faq.md
In several cases where the `faq.md` had **XXX**'s, we replaced the
text with equivalent text for someone to read (such as the **IOCCC
registration procedure** or **IOCCC submit server** is not ready).

In a number of cases, an HTML comment with a **XXX** inside,
was added as an indicator for the need of a future edit.

Added `make sitemap` as an alias for `make gen_sitemap`.

Performed `make gen_top_html sitemap` to do perform the above.