Add warning comment to 2013/dlowe/try.sh
It is important that this script does NOT prompt the user for each run
as this will break diff.sh. It might have been good if another script
allowed this but issue 3 is drawing to a close (as we all want) and it's
not that big of a deal anyway.