Improve bash version notes in bin/*.sh
The wording in the comments and the error message (in particular for
macOS) should now be clearer. Now in both places it shows how to install
bash in Homebrew. In commands I surrounded them like what was already
used, with " rather than 's, to make it consistent. There were other
cosmetic consistency changes as well.

Except in paths 'homebrew' was changed to 'Homebrew'.

In bin/output-year-index.sh there was a typo fix: it said to use it one
typically would do:

    -a bin/output-year-index.sh

which obviously is not correct.