Fix bug statuses of 2000/dlowe
The doesn't work with some platforms is wrong. It's the perl version and
the author noted this.

There is however a bug that was noted in the bugs file that was not in
the README.md file. This has been moved in the bugs.md file and
README.md file to above the INABIAF part.