Fix 2004/sds/try.sh for README.md changes
Commit d4cb036d776f361f239eaca17cd74cf040e93123 which edited
2004/sds/README.md broke the functionality of one of the features of the
code so the previous revision of README.md was copied to README_sds.md
and the script now uses that file.

This does mean the manifest is out of date and that will have to be
fixed another time but now the script works okay. Without this one of
the files to be compiled does not compile at all and it seems better
this way than to tamper with the code especially given what the entry
does and because it would take more time to fix anyway.