Updates to 2020/ferguson[12] wrt recipe
In 2020/ferguson1 the chocolate-cake.md file has been updated with a
suggestion that those who try it let me know how they like it and
sharing pictures (on mastodon).

In 2020/ferguson2 all chocolate cake related files have been removed. A
new file encrypted.html has been added with a new task one might wish to
try since the cake recipe is in 2020/ferguson1.

The make clobber rule in 2020/ferguson2 removes decrypted.html (the
recode.md file now shows how one might decrypt the encrypted.html in the
file decrypted.html but this file should not be committed and is now in,
along with a few others, .gitignore).

The html files have NOT been updated as manifest issues have to be
resolved first.