Fixes and improvements in 1980s index.html files
A point of interest is that in 1987 the README.md referred to Mark R
Horton but it is now Mary Ann Horton so I changed it to that. Also as
she has won I linked to not the Amazon author page but rather the
authors.html link. It's important to note though that the author page on
Amazon says Mark still, perhaps because that is how the book was
published (I guess - I didn't look). Still it seems right to change it
to how she identifies, rather than not.

Links were added for entries where referred to. Links were added to
results where not linked to. It is unclear if links to USENET groups
should be kept (besides the ones where the direct posts are known) or
removed so I have not added any that have missing links even those that
were already in other README.md files.

What remains to be done for the 1980s README.md files depends on whether
the USENET group links that are there should be removed or if those
missing should be added.

The rendered html files were tested on my web server and all appears
okay.