More fixes in 1980s README.md files
Where there is not a direct post when referring to a USENET group (for
results etc.) it is no longer a link to the group itself but just made
italic. That means only 1984 has it and only for the contest
announcement.

In 1989 the point about the Makefiles for fubar and tromp have been
clarified a bit. They were ambiguous before: it could have been read as
the Makefile for ALL entries but it was only for that entry. Also since
these points are about the Makefile (and in particular portability of
the original or alt versions) I moved the paragraph to right under the
part about running make to compile (this part has some additional fixes
that will be applied later on).

The years: 1984, 1985, 1986, 1987 and 1989 were updated. There was
nothing to do (for this commit) for 1988 but there is a change that will
be made after discussion.