Improve thanks-for-fixes up through 1992
For each YYYY header it is a link to the respective index.html page.

For each entry the index.html file for the entry is in the section name
of the entry (it used to not be a link there but rather in another line
with the source code on the same line). The source code is in a
subsection (that is the index.html is '#' and source code is '##') below
the index.html link.

(The rest of the years might happen later today or it might not happen
until another day due to a timing conflict here.)