add checkmark to the narbar of all constructed HTML
We added a checkmark all narbar's that has the make effect of
clicking on the **Nu HTML check this web page** link at the very
bottom.  This makes it much more convenient when one needs
to quickly check the HTML of a number of web pages.

Run `make www` to perform the above.