fix HTML problems related to pandoc and markdown image notation
Improve `make clobber` for `2004/anonymous`.

Instead of using this markdown element to embed an image:

```
![alt text](filename.png "Title")
```

Use a "img" HTML element with "alt", "alt", "width" and "length"
attributes:

```
<img src="filename.png"
 alt="describe the filename.png image for someone who cannot view it"
 width=PIXEL_WIDTH height=PIXEL_HEIGHT>
```

This was performed for:

```
2004/anonymous/index.html
2005/boutines/index.html
2011/goren/index.html
2012/vik/index.html
2013/birken/index.html
2013/cable1/index.html
2013/cable3/index.html
2013/hou/doc/example.html
2013/hou/index.html
2020/ferguson1/chocolate-cake.html
```

In the case of `2013/cable1/index.html` we reduce the displayed
image size by a factor for 2.  There is **NO** need to downsize the image
as this image size reduction produces a clean and sharp looking image.

In the case of `2013/hou/index.html` we reduce the displayed
image size by a factor for 2.  There is **NO** need to downsize the image
as this image size reduction produces a clean and sharp looking image.

In the case of `2020/ferguson1/chocolate-cake.html` we reduce the displayed
image size by a factor for 4.  There is **NO** need to downsize the image
as this image size reduction produces a clean and sharp looking image.

Run `make www` to perform the above.