Fix 2013/hou/Makefile for a.c
I noticed a problem with my improvement to the Makefile allowing one to
not use a temporary file so although the 'all' rule uses that method one
can still generate a.c so one can follow the author's remarks.

Links to this file were changed to code blocks, however, as the file
does not always exist. make clobber will remove a and a.c.