Several manifest updates
Changed 'description of entry in JSON' to 'entry summary and manifest in
JSON'.

Changed 'how to run' (for try scripts) to 'script to try entry'.

Changed 'how to run if you have the primes utility' to 'script to try
entry if you have the primes utility'.

Changed 'how to run alternative code' to 'script to try alternative
code'.

A number of other more specific (to particular entries or maybe even a
single entry, 2019/ciura if not any others) descriptions were also
changed.

Rebuilt index.html and .entry.json files.