Consistency and other fixes in manifest
As the index.html/README.md files say 'Alternate code' the 'alternative
source' has been changed to 'alternate code'. Similarly the text 'script
to try alternative code' was changed to 'script to try alternate code'.

A mistake was made where something referred to as being the alt code was
not the alt code (or the other way round).

Changed 'alternate alternative code' (or something like that) to 'second
alternate code'. This happened in steps and it might not, now that I
think on it, be the best wording, but this can be addressed later.

Rebuilt html files and .entry.json files.