Add issue form for website problems
The content can certainly be improved or added to. I still have to come
up with a form for other problems. It is not clear to me that an issue
form for bugs should be here as the code is not actually maintained and
one can see bugs.md for more details on this subject.