Fixes in 1992/adrian/try.sh
It did not show the use of the files submitted by the author, the
adrian.grep files that were not adrian.grep.try (though some do not show
output with the README.md file).

Split very long file (command shown) so that it shows it on two lines.

These were discovered when looking at the manifest (some changes made so
far but not ready for a commit).