Slight improvement to bin/gen-status.sh
Rather than do 'stat' or 'ls' do type -P on the tools and check that
they exist (or that the variables are not empty), exiting with a fatal
error (and exit code 5) if either are empty. The variables are used in
the script instead of the raw or literal 'stat' and 'ls'.