Add back bug status to 2001/cheong
Once upon a time after I had put in the bugs file that there is no arg
check and it was a feature, not a bug, it was changed to bug status.
This was a very simple fix that I then made but it's not necessary
either so I have undone it to make it more like the original. Now the
only difference from the original is my fix so that it compiles with
clang. This means the bugs file was also updated as was the index.html
file and the thanks file (which also has some other fixes that are
unrelated).