Restore bug status to 2001/ollinger
Once upon a time after adding the entry to the bugs file it was changed
so that it should be a bug to be fixed and when verifying things since
it was very quick to fix I added it. It now has been restored to the
original.

The bugs file was updated and the thanks file was updated accordingly
along with the index.html and README.md files.