improve bottom of index.html and FAQ 0.0
Improved information on how to view the status of the IOCCC, directing
people to `status.html` (which includes a copy of the contents of
`status.json`) instead of directly referring people to `status.json`.

Improved format of "FAQ 0.0: How may I submit to the IOCCC?".

Improved bottom of `index.html` to refer to the above.