Add note about top prio issue 2006 to news
It might be useful to know that the majority of the files should be
relatively quick to review and that of the three slower files one is
almost done.

I did not change the date of the news update but this seems like a
useful change (though maybe better wording could be used). If the date
is changed it would allow to update the status.json and status.html file
news date.