Rename 2012/zeitak/test.sh try.sh
This is because it was used in the same way as the try scripts and it
matches better other entries and looks better with the manifest (makes
it easier to see what it is for).

Rebuilt manifest, generated .entry.json and index.html files.

The thanks file was updated accordingly.