Restore web site to 2011/dlowe training data
Since GitHub has a problem with showing really big commits it seemed
like my attempts at changing 'web site' to 'website' were incomplete but
they were not. Something I noticed the first time shortly before I
closed that commit (as it seemed to be a problem) that I ended up
changing the file:

	2011/dlowe/dlowe-aux-data/english-1/Wireless_Days_2011_CFP.txt

as well. The problem is that this was a file used to train the neural
network of the entry so it should not have the change made.