Fix broken manifest links in 2011/dlowe
This is a fun one because the scripts that are used to update the
manifest did not report any problem because the problematic file (though
see below as I did rename another file too for very similar rationale,
also discussed below) actually exists on file system but when linked to
from a web server it does not exist. This is because the filename had
the html encodings for [ and ] (again see below) which the web server
interpreted as [ and ] but the file doesn't exist on the filesystem in
that form: instead it had the html encodings for those characters. This
file was

	2011/dlowe/dlowe-aux-data/english-0/CIR_APE_08_doc%5B1%5D.txt

and it has been renamed to

	2011/dlowe/dlowe-aux-data/english-0/CIR_APE_08_doc.txt

The other file h ad the literal [ and ] but as this is also not POSIX
safe the [ and ] plus what was in between it was removed. Thus the file

	2011/dlowe/dlowe-aux-data/png-1/image_thumb[40].png

was renamed to just

	2011/dlowe/dlowe-aux-data/png-1/image_thumb.png

Updated the manifest and regenerated the .entry.json and index.html
files.

As a side effect of the manifest and rebuild of the .entry.json and
index.html file two other entries were touched that somehow were not
done when I changed 'Best of show' to 'Best of Show' even though a git
diff will show the same thing (seemingly of course):

	-    "award" : "Best Of Show",
	+    "award" : "Best of Show",

plus a diff in their index.html files.

Those entries are 1991/brnstnd and 2001/jason though I am not sure why
as I'm pretty sure other entries also had the same change made and they
did not show up here.

Nevertheless that problem should now be resolved and the links in
2011/dlowe manifest now work (at least those two have been checked for
validity: others will be checked when I verify format and links in that
entry which will not be done until after the top level html files and
the earlier years).